Renamed GuiConfig to GuiConfiguration.
